Updated a couple of things :D
