+++ utils/__init__.py
@@ -1,8 +1,6 @@
 from .converters import *
 from .extra import *
 from .menus import *
-from .buttons import *
 from .embeds import *
 from .checks import *
-from .selects import *
-from .paginators import *
+from .views import *
